+++ src/include/sll/array.h
@@ -18,11 +18,23 @@ __SLL_FUNC void sll_array_clone(const sll_array_t* a,sll_array_t* o);
 
 
 
+__SLL_FUNC sll_array_length_t sll_array_count(const sll_array_t* a,sll_runtime_object_t* v);
+
+
+
+__SLL_FUNC sll_array_length_t sll_array_count_multiple(const sll_array_t* a,sll_runtime_object_t** v,sll_array_length_t vl);
+
+
+
 __SLL_FUNC void sll_array_create(sll_array_length_t l,sll_array_t* o);
 
 
 
 __SLL_FUNC void sll_array_join(const sll_array_t* a,const sll_array_t* b,sll_array_t* o);
+
+
+
+__SLL_FUNC sll_runtime_object_t* sll_array_pop(const sll_array_t* a,sll_array_t* o);
 
 
 
@@ -39,6 +51,14 @@ __SLL_FUNC void sll_array_remove_multiple(const sll_array_t* a,sll_runtime_objec
 
 
 __SLL_FUNC void sll_array_resize(const sll_array_t* a,sll_integer_t v,sll_array_t* o);
+
+
+
+__SLL_FUNC sll_runtime_object_t* sll_array_shift(const sll_array_t* a,sll_array_t* o);
+
+
+
+__SLL_FUNC void sll_array_unshift(const sll_array_t* a,sll_runtime_object_t* v,sll_array_t* o);