+++ src/include/sll/map.h
@@ -34,6 +34,10 @@ __SLL_FUNC void sll_map_keys(const sll_map_t* m,sll_array_t* o);
 
 
 
+__SLL_FUNC sll_runtime_object_t* sll_map_remove(const sll_map_t* m,sll_runtime_object_t* k,sll_map_t* o);
+
+
+
 __SLL_FUNC void sll_map_values(const sll_map_t* m,sll_array_t* o);