+++ src/include/sll/string.h
@@ -42,6 +42,10 @@ __SLL_FUNC void sll_string_clone(const sll_string_t* s,sll_string_t* d);
 
 
 
+__SLL_FUNC sll_string_length_t sll_string_count(const sll_string_t* a,const sll_string_t* b);
+
+
+
 __SLL_FUNC void sll_string_create(sll_string_length_t l,sll_string_t* o);
 
 
@@ -63,6 +67,18 @@ __SLL_FUNC sll_string_length_t sll_string_length(const sll_char_t* s);
 
 
 __SLL_FUNC void sll_string_lower_case(const sll_string_t* a,sll_string_t* o);
+
+
+
+__SLL_FUNC void sll_string_remove(const sll_string_t* a,const sll_string_t* s,sll_string_t* o);
+
+
+
+__SLL_FUNC void sll_string_replace(const sll_string_t* a,const sll_string_t* k,const sll_string_t* v,sll_string_t* o);
+
+
+
+__SLL_FUNC void sll_string_title_case(const sll_string_t* a,sll_string_t* o);