+++ src/sll/array.c
@@ -25,6 +25,33 @@ __SLL_FUNC void sll_array_clone(const sll_array_t* a,sll_array_t* o){
 
 
 
+__SLL_FUNC sll_array_length_t sll_array_count(const sll_array_t* a,sll_runtime_object_t* v){
+	sll_array_length_t o=0;
+	for (sll_array_length_t i=0;i<a->l;i++){
+		if (sll_operator_equal(a->v[i],v)){
+			o++;
+		}
+	}
+	return o;
+}
+
+
+
+__SLL_FUNC sll_array_length_t sll_array_count_multiple(const sll_array_t* a,sll_runtime_object_t** v,sll_array_length_t vl){
+	sll_array_length_t o=0;
+	for (sll_array_length_t i=0;i<a->l;i++){
+		for (sll_array_length_t j=0;j<vl;j++){
+			if (sll_operator_equal(a->v[i],*(v+j))){
+				o++;
+				break;
+			}
+		}
+	}
+	return o;
+}
+
+
+
 __SLL_FUNC void sll_array_create(sll_array_length_t l,sll_array_t* o){
 	if (!l){
 		SLL_ZERO_ARRAY(o);
@@ -48,6 +75,28 @@ __SLL_FUNC void sll_array_join(const sll_array_t* a,const sll_array_t* b,sll_arr
 	for (sll_array_length_t i=0;i<o->l;i++){
 		SLL_ACQUIRE(o->v[i]);
 	}
+}
+
+
+
+__SLL_FUNC sll_runtime_object_t* sll_array_pop(const sll_array_t* a,sll_array_t* o){
+	if (!a->l){
+		SLL_ZERO_ARRAY(o);
+		return SLL_ACQUIRE_STATIC_INT(0);
+	}
+	SLL_ACQUIRE(a->v[a->l-1]);
+	if (a->l==1){
+		SLL_ZERO_ARRAY(o);
+	}
+	else{
+		o->l=a->l-1;
+		o->v=malloc(o->l*sizeof(sll_runtime_object_t*));
+		memcpy(o->v,a->v,o->l*sizeof(sll_runtime_object_t*));
+		for (sll_array_length_t i=0;i<o->l;i++){
+			SLL_ACQUIRE(o->v[i]);
+		}
+	}
+	return a->v[a->l-1];
 }
 
 
@@ -146,5 +195,39 @@ __SLL_FUNC void sll_array_resize(const sll_array_t* a,sll_integer_t v,sll_array_
 	memcpy(o->v+v,a->v,a->l*sizeof(sll_runtime_object_t*));
 	for (sll_array_length_t i=0;i<a->l;i++){
 		SLL_ACQUIRE(a->v[i]);
+	}
+}
+
+
+
+__SLL_FUNC sll_runtime_object_t* sll_array_shift(const sll_array_t* a,sll_array_t* o){
+	if (!a->l){
+		SLL_ZERO_ARRAY(o);
+		return SLL_ACQUIRE_STATIC_INT(0);
+	}
+	SLL_ACQUIRE(a->v[0]);
+	if (a->l==1){
+		SLL_ZERO_ARRAY(o);
+	}
+	else{
+		o->l=a->l-1;
+		o->v=malloc(o->l*sizeof(sll_runtime_object_t*));
+		memcpy(o->v,a->v+1,o->l*sizeof(sll_runtime_object_t*));
+		for (sll_array_length_t i=0;i<o->l;i++){
+			SLL_ACQUIRE(o->v[i]);
+		}
+	}
+	return a->v[0];
+}
+
+
+
+__SLL_FUNC void sll_array_unshift(const sll_array_t* a,sll_runtime_object_t* v,sll_array_t* o){
+	o->l=a->l+1;
+	o->v=malloc(o->l*sizeof(sll_runtime_object_t*));
+	memcpy(o->v+1,a->v,a->l*sizeof(sll_runtime_object_t*));
+	o->v[0]=v;
+	for (sll_array_length_t i=0;i<o->l;i++){
+		SLL_ACQUIRE(o->v[i]);
 	}
 }