+++ src/sll/map.c
@@ -3,6 +3,8 @@
 #include <sll/common.h>
 #include <sll/gc.h>
 #include <sll/map.h>
+#include <sll/operator.h>
+#include <sll/static_object.h>
 #include <sll/types.h>
 #include <stdlib.h>
 #include <string.h>
@@ -75,6 +77,33 @@ __SLL_FUNC void sll_map_keys(const sll_map_t* m,sll_array_t* o){
 		o->v[i]=m->v[i<<1];
 		SLL_ACQUIRE(o->v[i]);
 	}
+}
+
+
+
+__SLL_FUNC sll_runtime_object_t* sll_map_remove(const sll_map_t* m,sll_runtime_object_t* k,sll_map_t* o){
+	for (sll_map_length_t i=0;i<m->l;i++){
+		SLL_ACQUIRE(m->v[(i<<1)+1]);
+		if (sll_operator_equal(m->v[i<<1],k)){
+			o->l=m->l-1;
+			o->v=malloc((o->l<<1)*sizeof(sll_runtime_object_t*));
+			memcpy(o->v,m->v,(i<<1)*sizeof(sll_runtime_object_t*));
+			memcpy(o->v+(i<<1),m->v+((i+1)<<1),((m->l-i-1)<<1)*sizeof(sll_runtime_object_t*));
+			sll_runtime_object_t* v=m->v[(i<<1)+1];
+			i++;
+			while (i<m->l){
+				SLL_ACQUIRE(m->v[i<<1]);
+				SLL_ACQUIRE(m->v[(i<<1)+1]);
+				i++;
+			}
+			return v;
+		}
+		SLL_ACQUIRE(m->v[i<<1]);
+	}
+	o->l=m->l;
+	o->v=malloc((o->l<<1)*sizeof(sll_runtime_object_t*));
+	memcpy(o->v,m->v,(o->l<<1)*sizeof(sll_runtime_object_t*));
+	return SLL_ACQUIRE_STATIC_INT(0);
 }